Refactor the types to only have the headers once
