+++ ddsrecorder_participants/include/ddsrecorder_participants/common/types/monitoring/ddsrecorder_status/v1/DdsRecorderMonitoringStatus.idl
@@ -1,4 +1,4 @@
-#include <ddspipe_core/types/monitoring/status/v2/MonitoringStatus.idl>
+#include <ddspipe_core/types/monitoring/status/v1/MonitoringStatus.idl>
 
 struct DdsRecorderMonitoringErrorStatus {
     boolean mcap_file_creation_failure;
