+++ ddsrecorder_participants/include/ddsrecorder_participants/common/types/monitoring/ddsrecorder_status/v1/DdsRecorderMonitoringStatusCdrAux.ipp
