+++ ddsrecorder_participants/include/ddsrecorder_participants/common/types/monitoring/ddsrecorder_status/v1/DdsRecorderMonitoringStatusPubSubTypes.h
@@ -31,7 +31,7 @@
 
 #include "DdsRecorderMonitoringStatus.h"
 
-#include "ddspipe_core/types/monitoring/status/v2/MonitoringStatusPubSubTypes.h"
+#include "ddspipe_core/types/monitoring/status/v1/MonitoringStatusPubSubTypes.h"
 
 #if !defined(GEN_API_VER) || (GEN_API_VER != 2)
 #error \
