+++ ddsrecorder_participants/src/cpp/common/types/monitoring/ddsrecorder_status/v2/DdsRecorderMonitoringStatus.cxx
@@ -26,7 +26,7 @@ char dummy;
 }  // namespace
 #endif  // _WIN32
 
-#include "DdsRecorderMonitoringStatus.h"
+#include <ddsrecorder_participants/common/types/monitoring/ddsrecorder_status/v2/DdsRecorderMonitoringStatus.h>
 
 #if FASTCDR_VERSION_MAJOR > 1
 
@@ -254,6 +254,6 @@ DdsRecorderMonitoringErrorStatus& DdsRecorderMonitoringStatus::ddsrecorder_error
 
 
 // Include auxiliary functions like for serializing/deserializing.
-#include "DdsRecorderMonitoringStatusCdrAux.ipp"
+#include <ddsrecorder_participants/common/types/monitoring/ddsrecorder_status/v1/DdsRecorderMonitoringStatusCdrAux.ipp>
 
 #endif // FASTCDR_VERSION_MAJOR > 1
