+++ ddsrecorder_participants/src/cpp/common/types/monitoring/ddsrecorder_status/v2/DdsRecorderMonitoringStatusv1.cxx
@@ -26,7 +26,7 @@ char dummy;
 }  // namespace
 #endif  // _WIN32
 
-#include "DdsRecorderMonitoringStatus.h"
+#include <ddsrecorder_participants/common/types/monitoring/ddsrecorder_status/v2/DdsRecorderMonitoringStatus.h>
 
 #if FASTCDR_VERSION_MAJOR == 1
